7. Sync Damage / Health

+++ Assets/Scripts/PlayerWeapon.cs
@@ -6,6 +6,6 @@ using UnityEngine;
 public class PlayerWeapon
 {
     public string Name = "Glock";
-    public float Damage = 10f;
+    public int Damage = 10;
     public float Range = 100f;
 }
